fix: correct PositionSide import in useSpotlight.ts and refactor helper to coordinate file to fix ESLint max-lines error
